[Priroda] Split debugger frontends into modules
