+++ priroda/src/frontend/mod.rs
@@ -1,0 +1,5 @@
+mod cli;
+mod dap;
+
+pub(super) use cli::Cli;
+pub(super) use dap::Dap;
